Parsing Model Entities
